adicionei opções para as tasks setup e build

+++ tasks/setup.sh
@@ -1,12 +1,38 @@
 #!/bin/bash
 set -e
 
+if [ $# -ne 2 ]; then
+    echo "Usage: $0 <simulation option> <parallel/sequential>"
+    exit 1
+fi
+
 vcs import < src/dependencies.repos src
 vcs import < src/frtl_2023/dependencies.repos src
-vcs import < src/simulation/dependencies.repos src
 
-sudo apt-get update
-sudo apt remove ros-humble-ros-gz-bridge
+if [ "$1" = "with simulation" ]
+then
+    vcs import < src/simulation/dependencies.repos src
+    sudo apt-get update
+    sudo apt remove ros-humble-ros-gz-bridge
+elif [ "$1" = "without simulation" ]
+then
+    sudo rm -r src/simulation
+    sudo apt-get update
+else
+    echo "Possible options: \"with simulation\", \"without simulation\""
+    exit 1
+fi
+
+if [ $2 = sequential ]
+then
+    SEQUENTIAL_BUILD="--executor sequential"
+elif [ $2 = parallel ]
+then
+    SEQUENTIAL_BUILD=""
+else
+    echo "Possible options: sequential, parallel"
+    exit 1
+fi
 
 source /opt/ros/humble/setup.bash
 rosdep update
@@ -18,4 +44,5 @@ colcon build \
         --symlink-install \
         --event-handlers console_direct+ \
         --cmake-args "-DCMAKE_BUILD_TYPE=$BUILD_TYPE" "-DCMAKE_EXPORT_COMPILE_COMMANDS=On" \
-        -Wall -Wextra -Wpedantic
+        -Wall -Wextra -Wpedantic \
+        $SEQUENTIAL_BUILD
